requirements.txt
django
psycopg2
black
Pillow
djangorestframework
django-rest-authtoken
dj-rest-auth
django-allauth
